Start fixing story contexts
