+++ apps/storybook/.storybook/main.ts
@@ -1,14 +1,14 @@
+// biome-ignore lint/correctness/noNodejsModules: This is a Node.js module
+import { dirname, join } from 'node:path';
 import type { StorybookConfig } from '@storybook/nextjs';
-
-import { dirname, join } from 'path';
 
 /**
  * This function is used to resolve the absolute path of a package.
  * It is needed in projects that use Yarn PnP or are set up within a monorepo.
  */
-function getAbsolutePath(value: string): any {
-  return dirname(require.resolve(join(value, 'package.json')));
-}
+const getAbsolutePath = (value: string) =>
+  dirname(require.resolve(join(value, 'package.json')));
+
 const config: StorybookConfig = {
   stories: [
     '../stories/**/*.mdx',
@@ -26,4 +26,5 @@ const config: StorybookConfig = {
   },
   staticDirs: ['../public'],
 };
+
 export default config;
