+++ apps/storybook/.storybook/preview.tsx
@@ -1,0 +1,29 @@
+import { Toaster } from '@repo/design-system/components/ui/sonner';
+import { TooltipProvider } from '@repo/design-system/components/ui/tooltip';
+import { ThemeProvider } from '@repo/design-system/providers/theme';
+import type { Preview } from '@storybook/react';
+
+import '@repo/design-system/styles/globals.css';
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <ThemeProvider>
+        <TooltipProvider>
+          <Story />
+        </TooltipProvider>
+        <Toaster />
+      </ThemeProvider>
+    ),
+  ],
+};
+
+export default preview;
